Split parcel_core crate
