+++ crates/parcel_core/src/lib.rs
@@ -1,8 +1,7 @@
 //! Core re-implementation in Rust
 
-/// napi versions of `crate::core::requests`
-mod js_requests;
+pub mod hash;
 /// New-type for paths relative to a project-root
-mod project_path;
+pub mod project_path;
 /// Request types and run functions
-mod requests;
+pub mod requests;
