+++ crates/parcel_core/src/requests/config_request/mod.rs
@@ -4,11 +4,11 @@
 //! file.
 use std::path::Path;
 
-use crate::core::project_path::ProjectPath;
 use napi_derive::napi;
 use parcel_resolver::FileSystem;
 
-use crate::core::requests::request_api::RequestApi;
+use crate::project_path::ProjectPath;
+use crate::requests::request_api::RequestApi;
 
 pub type InternalGlob = String;
 
@@ -154,11 +154,12 @@ struct RequestOptions {}
 
 #[cfg(test)]
 mod test {
-  use super::*;
-  use crate::core::requests::config_request::run_config_request;
-  use crate::core::requests::request_api::MockRequestApi;
   use parcel_filesystem::in_memory_file_system::InMemoryFileSystem;
   use parcel_filesystem::os_file_system::OsFileSystem;
+
+  use super::*;
+  use crate::requests::config_request::run_config_request;
+  use crate::requests::request_api::MockRequestApi;
 
   #[test]
   fn test_run_empty_config_request_does_nothing() {
