+++ crates/parcel_core/src/requests/request_api/js_request_api.rs
@@ -6,9 +6,9 @@ use napi::JsObject;
 use napi::JsUnknown;
 use parcel_napi_helpers::call_method;
 
-use crate::core::requests::config_request::InternalFileCreateInvalidation;
-use crate::core::requests::request_api::RequestApi;
-use crate::core::requests::request_api::RequestApiResult;
+use crate::requests::config_request::InternalFileCreateInvalidation;
+use crate::requests::request_api::RequestApi;
+use crate::requests::request_api::RequestApiResult;
 
 /// This is a "delegate" implementation of `RequestApi` that delegates calls to a
 /// JavaScript object.
